# refactor(ansi-to-html): Avoid peeking while parsing ANSI codes