--- a/crates/ansi-to-html/src/ansi/parse.rs
+++ b/crates/ansi-to-html/src/ansi/parse.rs
@@ -1,15 +1,57 @@
-use std::{ops::Range, str::CharIndices};
+use std::{mem, ops::Range, str::CharIndices};
+
+const ESCAPE: char = '\u{1b}';
 
 #[must_use]
 pub(crate) struct AnsiParser<'text> {
     text: &'text str,
     chars: CharIndices<'text>,
+    /// Stores whether the last iteration of `.next()` ended on an escape character
+    ///
+    /// The parser sifts through text to find ANSI codes which **always** start with an escape
+    /// character. We can exploit this to speed up the hot path of iterating over plain text by
+    /// unconditionally iterating over characters and storing whether the character that broke us
+    /// out was an escape in this field. This avoids the need to peek each character to see if it's
+    /// the start of an ANSI code by re-framing the ansi code parsing to start _after_ the intial
+    /// escape and adjusting the starting index to still include it
+    was_on_esc: bool,
 }
 
 impl<'text> AnsiParser<'text> {
     pub(crate) fn new(text: &'text str) -> Self {
         let chars = text.char_indices();
-        Self { text, chars }
+        let was_on_esc = false;
+        Self {
+            text,
+            chars,
+            was_on_esc,
+        }
+    }
+
+    fn parse_ansi_code_after_esc(&mut self) -> Option<AnsiFragment<'text>> {
+        let mut fsm = AnsiFsm::new(&mut self.chars);
+        loop {
+            let Some(status) = fsm.peek() else {
+                break Some(AnsiFragment::Text(&self.text[fsm.span()]));
+            };
+
+            match status {
+                Status::InSequence => _ = fsm.next(),
+                Status::Accept => {
+                    fsm.next();
+                    break Some(AnsiFragment::Sequence(&self.text[fsm.span()]));
+                }
+                // NOTE(cosmic): niche case, but behavior is diverging from the regex here
+                // which would return invalid ansi codes _along with_ any surround text whereas
+                // we're emitting them separately atm
+                Status::RejectAsText => {
+                    // Fortunately the starting ESC of an ANSI sequence can't appear within
+                    // the sequence itself, so we don't need to worry about any backtracking or
+                    // reparsing
+                    break Some(AnsiFragment::Text(&self.text[fsm.span()]));
+                }
+            }
+        }
     }
 }
 
@@ -23,61 +65,47 @@ impl<'text> Iterator for AnsiParser<'text> {
     type Item = AnsiFragment<'text>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let mut dfa = Dfa::new(&mut self.chars).unwrap();
-
-        match dfa.next()? {
-            // Walk through the current run of text
-            Status::Waiting => {
-                while dfa.peek() == Some(Status::Waiting) {
-                    dfa.next();
-                }
-                Some(AnsiFragment::Text(&self.text[dfa.span()]))
+        if mem::take(&mut self.was_on_esc) {
+            self.parse_ansi_code_after_esc()
+        } else {
+            let start_idx = self.chars.offset();
+            let (_, c) = self.chars.next()?;
+            if c == ESCAPE {
+                return self.parse_ansi_code_after_esc();
             }
-            // Walk through the ansi sequence
-            Status::InSequence => loop {
-                let Some(status) = dfa.peek() else {
-                    break Some(AnsiFragment::Text(&self.text[dfa.span()]));
-                };
-
-                match status {
-                    Status::Waiting => unreachable!("We're already past `Waiting`"),
-                    Status::InSequence => _ = dfa.next(),
-                    Status::Accept => {
-                        dfa.next();
-                        break Some(AnsiFragment::Sequence(&self.text[dfa.span()]));
-                    }
-                    // TODO(cosmic): niche case, but behavior is diverging from the regex here
-                    // which would return invalid ansi codes _along with_ any surround text whereas
-                    // we're emitting them separately atm
-                    Status::RejectAsText => {
-                        // Fortunately the starting ESC of an ANSI sequence can't appear within
-                        // the sequence itself, so we don't need to worry about any backtracking or
-                        // reparsing
-                        break Some(AnsiFragment::Text(&self.text[dfa.span()]));
-                    }
+            while let Some((_, c)) = self.chars.next() {
+                if c == ESCAPE {
+                    self.was_on_esc = true;
+                    break;
                 }
-            },
-            _ => unreachable!("No other possible statuses after just one char"),
+            }
+            let end_offset = if self.was_on_esc {
+                ESCAPE.len_utf8()
+            } else {
+                0
+            };
+            let end_idx = self.chars.offset() - end_offset;
+            Some(AnsiFragment::Text(&self.text[start_idx..end_idx]))
         }
     }
 }
 
-/// A small DFA that detects and emits ANSI codes from an underlying `CharIndices`
-struct Dfa<'short, 'text: 'short> {
+/// A small finite state machine that parses ANSI codes after the initial ESC
+struct AnsiFsm<'short, 'text: 'short> {
     chars: &'short mut CharIndices<'text>,
     state: State,
     start_idx: usize,
 }
 
-impl<'short, 'text> Dfa<'short, 'text> {
-    fn new(chars: &'short mut CharIndices<'text>) -> Option<Self> {
+impl<'short, 'text> AnsiFsm<'short, 'text> {
+    fn new(chars: &'short mut CharIndices<'text>) -> Self {
         let state = State::default();
-        let start_idx = chars.offset();
-        Some(Self {
+        let start_idx = chars.offset() - ESCAPE.len_utf8();
+        Self {
             chars,
             state,
             start_idx,
-        })
+        }
     }
 
     fn peek(&self) -> Option<Status> {
@@ -101,12 +129,10 @@ impl<'short, 'text> Dfa<'short, 'text> {
 
 #[derive(Clone, Copy, Default)]
 enum State {
-    /// Waiting to see an escape
     #[default]
-    Init,
+    Escape,
     Trap,
     Accept,
-    Escape,
     EscapeOpenParen,
     /// Control Sequence Introducer (CSI) - Indicates the start of an ansi sequence
     Csi,
@@ -117,9 +143,6 @@ enum State {
 impl State {
     fn munch(&mut self, c: char) {
         *self = match (*self, c) {
-            // Waiting for ESC
-            (Self::Init, '\u{1b}') => Self::Escape,
-            (Self::Init, _) => Self::Init,
             // Weird `<ESC>(B` ansi code
             (Self::Escape, '(') => Self::EscapeOpenParen,
             (Self::EscapeOpenParen, 'B') => Self::Accept,
@@ -139,7 +162,6 @@ impl State {
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 enum Status {
-    Waiting,
     InSequence,
     Accept,
     RejectAsText,
@@ -148,7 +170,6 @@ enum Status {
 impl From<State> for Status {
     fn from(state: State) -> Self {
         match state {
-            State::Init => Self::Waiting,
             State::Trap => Self::RejectAsText,
             State::Accept => Self::Accept,
             State::Escape
@@ -163,6 +184,19 @@ impl From<State> for Status {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn plain() {
+        let parser = AnsiParser::new("Hello World!");
+        let fragments: Vec<_> = parser.into_iter().collect();
+        insta::assert_debug_snapshot!(fragments, @r#"
+        [
+            Text(
+                "Hello World!",
+            ),
+        ]
+        "#);
+    }
 
     #[test]
     fn variety() {