This PR follows through with the optimization mentioned in https://github.com/Aloso/to-html/pull/45#discussion_r1912527412 where we can avoid peeking entirely when looking for the initial escape that starts an ANSI code by storing whether an escape character broke us out from a run of plain text

This **significantly** improves `ansi-to-html`'s performance across the board on #49's benchmarks. Mean times when running benches for 3 seconds for instance

|Bench | `main` | https://github.com/Aloso/to-html/pull/52/commits/438e13ad7abde565bd0fd0f53b65a43b5ac96ae3 alone | this PR |
| --: | :--: | :--: | :--: |
| (AnsiHeavy, Esc, Opt) | 63.94 MB/s | 82.22 MB/s | 92.55 MB/s |
| (AnsiHeavy, Esc, SkipOpt) | 76.54 MB/s | 104.3 MB/s | 122.1 MB/s |
| (AnsiHeavy, SkipEsc, Opt) | 89.43 MB/s | 127.8 MB/s | 148.5 MB/s |
| (AnsiHeavy, SkipEsc, SkipOpt) | 107.4 MB/s | 168.3 MB/s | 203 MB/s |
| (PlainText, Esc, Opt) | 108.6 MB/s | 215.7 MB/s | 228 MB/s |
| (PlainText, Esc, SkipOpt) | 109.6 MB/s | 217.5 MB/s | 228.8 MB/s |
| (PlainText, SkipEsc, Opt) | 182.5 MB/s| 1.005 GB/s | 1.007 GB/s |
| (PlainText, SkipEsc, SkipOpt) | 183 MB/s | 1.066 GB/s | 1.08 GB/s |