@CosmicHorrorDev 2025-02-11
Went ahead and followed through with generalizing this approach to retain the character that broke us out of our current token being parsed to avoid peeking entirely. Benchmarks show good improvements over just https://github.com/Aloso/to-html/pull/52/commits/438e13ad7abde565bd0fd0f53b65a43b5ac96ae3

```
Timer precision: 12 ns
convert                              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ convert                                         │               │               │               │         │
   ├─ (AnsiHeavy, Esc, Opt)          1.472 ms      │ 2.251 ms      │ 1.494 ms      │ 1.502 ms      │ 1975    │ 1975
   │                                 94.4 MB/s     │ 61.75 MB/s    │ 93.01 MB/s    │ 92.55 MB/s    │         │
   ├─ (AnsiHeavy, Esc, SkipOpt)      1.109 ms      │ 1.443 ms      │ 1.13 ms       │ 1.137 ms      │ 2637    │ 2637
   │                                 125.2 MB/s    │ 96.31 MB/s    │ 122.9 MB/s    │ 122.1 MB/s    │         │
   ├─ (AnsiHeavy, SkipEsc, Opt)      913.3 µs      │ 1.497 ms      │ 931.9 µs      │ 935.7 µs      │ 3206    │ 3206
   │                                 152.2 MB/s    │ 92.86 MB/s    │ 149.1 MB/s    │ 148.5 MB/s    │         │
   ├─ (AnsiHeavy, SkipEsc, SkipOpt)  659.6 µs      │ 1.076 ms      │ 682.2 µs      │ 684.6 µs      │ 4381    │ 4381
   │                                 210.7 MB/s    │ 129.1 MB/s    │ 203.7 MB/s    │ 203 MB/s      │         │
   ├─ (PlainText, Esc, Opt)          574.1 µs      │ 806.7 µs      │ 607.3 µs      │ 609.6 µs      │ 4920    │ 4920
   │                                 242.1 MB/s    │ 172.3 MB/s    │ 228.9 MB/s    │ 228 MB/s      │         │
   ├─ (PlainText, Esc, SkipOpt)      577.3 µs      │ 681.6 µs      │ 605.9 µs      │ 607.6 µs      │ 4937    │ 4937
   │                                 240.7 MB/s    │ 203.9 MB/s    │ 229.4 MB/s    │ 228.8 MB/s    │         │
   ├─ (PlainText, SkipEsc, Opt)      136.4 µs      │ 224.9 µs      │ 137.4 µs      │ 137.9 µs      │ 21720   │ 21720
   │                                 1.018 GB/s    │ 618.1 MB/s    │ 1.011 GB/s    │ 1.007 GB/s    │         │
   ╰─ (PlainText, SkipEsc, SkipOpt)  127.7 µs      │ 184.3 µs      │ 128.2 µs      │ 128.7 µs      │ 23280   │ 23280
                                     1.088 GB/s    │ 754.1 MB/s    │ 1.084 GB/s    │ 1.08 GB/s     │         │
```